+++ src/indicators/mod.rs
@@ -227,34 +227,4 @@ impl CandleProcessor for dyn Indicator {
 
         Ok(())
     }
-
-    /// Process new candle data
-    ///
-    /// This is the main interface for processing new candle data. It is meant to be called with
-    /// new candle data as it is received from the market.
-    ///
-    /// # Arguments
-    /// * `candles` - New candle data. Should be larger than processing window.
-    ///
-    /// # Panics
-    /// * If the DataFrame does not contain more than one row
-    fn process_new_candles(&mut self, candles: &DataFrame) -> Result<(), Self::ErrorType> {
-        assert!(
-            candles.height() > 1,
-            "DataFrame must contain more than one row"
-        );
-
-        match self.process_graph(candles) {
-            Ok(_) => {}
-            Err(e) => return Err(IndicatorProcessingError::GraphError(e)),
-        };
-
-        match self.process_signals(candles) {
-            Ok(_) => {}
-            Err(e) => return Err(IndicatorProcessingError::SignalError(e)),
-        }
-
-        Ok(())
-    }
-
 }